Adding the Housing VS Stocks python code
This is the weapon to use against people that say "buying a house is a good investment compared to renting". With the opportunity to invest money by simply tracking the S&P500 you will on average make more by renting and then buying a home. There are a few generalizations made for simplicity, but the main calculations are done.